requirements.txt
mlflow>=2.17.0
numpy>=1.26.0
pandas>=2.2.0
opencv-python>=4.10.0
Pillow>=10.4.0
tqdm>=4.66.0
pyiqa>=0.1.14
pypiqe>=1.2.0
